MinimalApiDemo/Calculators/MailCalculator.cs
using MinimalApiDemo.Calculators.Interfaces;
using MinimalApiDemo.RequestModels;
using System.Globalization;
using System.Text.RegularExpressions;

#pragma warning disable CS0168 // Variable is declared but never used
namespace MinimalApiDemo.Calculators
{
    internal class MailCalculator : IMailCalculator
    {
        // taken from https://learn.microsoft.com/en-us/dotnet/standard/base-types/how-to-verify-that-strings-are-in-valid-email-format?redirectedfrom=MSDN
        public bool CalculateIsValidMail(TextModel model)
        {
            string mail = model.Text;

            if (string.IsNullOrWhiteSpace(mail))
            {
                return false;
            }

            try
            {
                // Normalize the domain
                mail = Regex.Replace(mail, @"(@)(.+)$", DomainMapper,
                                      RegexOptions.None, TimeSpan.FromMilliseconds(200));

                // Examines the domain part of the email and normalizes it.
                static string DomainMapper(Match match)
                {
                    // Use IdnMapping class to convert Unicode domain names.
                    IdnMapping idn = new IdnMapping();

                    // Pull out and process domain name (throws ArgumentException on invalid)
                    string domainName = idn.GetAscii(match.Groups[2].Value);

                    return match.Groups[1].Value + domainName;
                }
            }
            catch (RegexMatchTimeoutException e)
            {
                return false;
            }
            catch (ArgumentException e)
            {
                return false;
            }

            try
            {
                return Regex.IsMatch(mail,
                    @"^[^@\s]+@[^@\s]+\.[^@\s]+$",
                    RegexOptions.IgnoreCase, TimeSpan.FromMilliseconds(250));
            }
            catch (RegexMatchTimeoutException)
            {
                return false;
            }
        }
    }
}
